requirements.txt
alpaca-py
pandas
python-dotenv
gspread
streamlit
